Create README.md
In Iran, due to the existing problems, it is not possible to confirm and verify the phone number by SMS. SMS will be sent, but even with the reduction of the sending amount, which is 20 times the price of sending a normal SMS in Iran, the cost will be reduced.  But it is not confirmed by the existing phone number and we receive an error message that the SMS was not delivered.  They have started mining in the hope that we will finally succeed soon to have a good income. The economic situation in Iran is really very bad, and if we succeed in buying and selling this currency code, it means we will achieve great success for all countries, especially the Iranian people.  With your help, of course, all the cryptocurrencies and all the coins, Bitcoin, Dodge Coins, and all of them can be easily traded in Iran, and we hope you friends will also help the Iranians to achieve their goal.  They have been waiting for success for several years and months, which I know with Moff plans  This is how it happens to you and with the help of God, all the problems are solved.  The world has traded and done economic work. We hope that the Ramzars, who have succeeded so far and can be easily paid in Iran, will ask you to help them solve the remaining small problems for this large number of families in Iran.  I typed the Rafarsi text and translated it into English with Google Translator.  I can somewhat moderate that I can not write or read English to a great extent and I use Google Translator. Thank you.